completed basic spring authorization server implementation

+++ src/main/resources/db/migration/V4__ADD_SESSIONSTATUS.sql
@@ -1,2 +1,2 @@
 ALTER TABLE session
-    ADD status SMALLINT NULL;
+    ADD status TINYINT NULL;
